refactor: navigator -> navigators 네이밍 변경
