+++ src/navigators/SignUpStackNavigator.tsx
